implement incomplete patching support
